Going to do more work tomorrow. Done for today
